ci: add coverage badge
Added a scripts/regenerate-badges.sh script that currently only generates a coverage badge based on the last run of pytest. The script is supposed to only be invoked as a commitizen pre-bump-hook - the respective configuration was added to pyproject.toml. The badge is referenced in the README.md and reflects the total percentage coverage of the source files (i.e. excluding tests) of the latest published version as represented by the PyPI version badge.